move StaticRouteService to static_route.py
